GCS_MAVLink: correct format in short-write panic message

+++ libraries/GCS_MAVLink/GCS_MAVLink.cpp
@@ -130,7 +130,7 @@ void comm_send_buffer(mavlink_channel_t chan, const uint8_t *buf, uint8_t len)
     const size_t written = mavlink_comm_port[chan]->write(buf, len);
 #if CONFIG_HAL_BOARD == HAL_BOARD_SITL
     if (written < len) {
-        AP_HAL::panic("Short write on UART: %lu < %u", written, len);
+        AP_HAL::panic("Short write on UART: %lu < %u", (unsigned long)written, len);
     }
 #else
     (void)written;
